requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
scikit-learn
h5py
pillow
python-multipart